rename a couple installer Util classes to Helper
