+++ src/lib/Zikula/Bundle/CoreInstallerBundle/Controller/UpgraderController.php
@@ -43,7 +43,7 @@ class UpgraderController extends AbstractController
         }
 
         // check php
-        $ini_warnings = $this->util->initPhp();
+        $ini_warnings = $this->controllerHelper->initPhp();
         if (count($ini_warnings) > 0) {
             $request->getSession()->getFlashBag()->add('warning', implode('<hr>', $ini_warnings));
         }
@@ -57,7 +57,7 @@ class UpgraderController extends AbstractController
         if ($currentStage instanceof WizardCompleteInterface) {
             return $currentStage->getResponse($request);
         }
-        $templateParams = $this->util->getTemplateGlobals($currentStage);
+        $templateParams = $this->controllerHelper->getTemplateGlobals($currentStage);
         $templateParams['headertemplate'] = '@ZikulaCoreInstaller/upgradeheader.html.twig';
         if ($wizard->isHalted()) {
             $request->getSession()->getFlashBag()->add('danger', $wizard->getWarning());
