+++ src/lib/Zikula/Bundle/CoreInstallerBundle/Stage/Install/CreateAdminStage.php
@@ -72,6 +72,6 @@ class CreateAdminStage implements StageInterface, FormHandlerInterface, InjectCo
 
     public function handleFormResult(FormInterface $form)
     {
-        $this->container->get('zikula_core_installer.controller.util')->writeEncodedAdminCredentials($this->yamlManager, $form->getData());
+        $this->container->get('zikula_core_installer.controller.helper')->writeEncodedAdminCredentials($this->yamlManager, $form->getData());
     }
 }
